update(app): restructure TabBar files and remove text from TabButton
