+++ packages/app/src/components/navigation/TabBar/TabBar.tsx
@@ -1,20 +1,18 @@
 import {type Component, type ComponentProps, For} from 'solid-js';
 
-import TabButton from '@/components/navigation/TabButton';
-
-import styles from './styles.module.scss';
+import TabButton from '@/components/navigation/TabBar/TabButton';
 
 const TabBar: Component<{
     links: ComponentProps<typeof TabButton>[];
 }> = props => {
     return (
         <div class="flex">
-            <nav class={styles.nav}>
+            <nav class="flex-1 flex h-14 bg-gray-900 p-2">
                 <For each={props.links}>
                     {link => (
                         <TabButton
+                            class="flex-1"
                             icon={link.icon}
-                            label={link.label}
                             route={link.route}
                         />
                     )}
