+++ packages/app/src/components/navigation/TabBar/index.ts
@@ -1,0 +1,3 @@
+import TabBar from './TabBar';
+
+export default TabBar;
