+++ packages/app/src/components/navigation/TabBar/styles.module.scss
@@ -1,10 +1,26 @@
-.nav {
-    flex: 1;
+.tab-button {
+    position: relative;
     display: flex;
+    justify-content: center;
+    align-items: center;
 
-    box-shadow: 0 -1px 5px 0 rgba(0, 0, 0, 0.4);
+    @apply text-gray-500;
 
-    & > * {
-        flex: 1;
+    &::after {
+        content: '';
+        position: absolute;
+        border-radius: 9999px;
+        width: 50%;
+        height: 100%;
+    }
+
+    &:hover::after,
+    &.tab-button--active::after {
+        @apply bg-white/10;
+    }
+
+    &:hover,
+    &.tab-button--active {
+        @apply text-primary;
     }
 }
